allow to read recipients from http headers (#16)
This PR allows to pass the --recipients-header option to the beta dashboard subcommand and read the recipients from a specified http header. This allows for example to place bpmon behind a keycloak-proxy (https://github.com/gambol99/keycloak-proxy) for authentification and use the generate 'X-Auth-Groups' header do grant access do BPs via 'recipients' in the BP definition.