Improve configuration infrastucture (#25)
Currently configuration parsing and validation is implemented very rudimentary. This pull request addresses this and implements a bunch of features:

Defaults can be defined and will be used by default by injecting an YAML section and anchor into the config content while parsing. This provides a reproducible way to set defaults for the user. The injected YAML can be printed via bpmon -c [file] config raw. Also a new flag -d/--defaults can be set to false in order to prevent the injection of defaults.
Validation can now be implemented quite easily via Config.Validate()
Comments from the fields of the Config struct and its children are used to generate documentation. The documentation as comments in the defaulted example configuration generated via bpmon config init